Create jtabs.html
A tab script that is NOT id based, which makes it more template friendly that the native jQuery UI tab script. Additionally, persistence is maintained through local storage and NOT through changing the hash, making the tab script more friendly for a web-app which may use hash change events for navigation purposes.